requirements.txt

tensorflow-gpu==1.5.0
keras==2.0.8
pillow
numpy